lesson_30/homework_30/requirements.txt
psycopg2-binary
pytest
allure-pytest